[Minor] Revert change in offline inference example
